requirements.txt
beautifulsoup4==4.12.2
Pillow==10.0.0
Requests==2.31.0
